SetupTests: Add requestAnimationFrame back (#7)
This update adds `window.requestAnimationFrame` mocking back to setupTests.
This was briefly removed to help integrate with the main application.
However, rAF mocking is very much a thing that is required for more complicated
React components (e.g. Popover, Dropdown, etc...).

The solution involved caching the mock implementation as a variable, to make
it more reliable to `mockRestore`. Also, to handle the case of the callback
being a `schedule` function, to prevent the maximum callstack issue.